backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]
supabase
gotrue
postgrest
httpx>=0.25.0
websockets>=13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
db-dtypes
pyarrow
cryptography

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Caching (optional, for distributed caching)
redis>=5.0.0